pyproject.toml
[project]
name = "marcapo-linkedin"
version = "0.1.0"
description = "LinkedIn-Microsoft Dynamics CRM synchronization system using MCP servers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0",
    "pydantic-ai>=0.2.0",
    "ollama>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
linkedin-sync = "sync.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_servers", "sync"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0", 
    "black>=24.0.0",
    "ruff>=0.1.0",
]
